requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
ultralytics==8.3.180
torch==2.8.0
torchvision==0.23.0
opencv-python==4.10.0.84
pillow==10.4.0
pyyaml==6.0.1
pydantic==1.10.14
transformers==4.44.0
segmentation_models_pytorch==0.3.4
